% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_song.R
\name{random_song}
\alias{random_song}
\title{Select a random song}
\usage{
random_song(artist = "Beyoncé", data = billboard::wiki_hot_100s)
}
\arguments{
\item{data}{The dataset to use; defaults to billboard::wiki_hot_100s}

\item{A}{band / artist to select a song from}
}
\value{
A random song gets printed to the console
}
\description{
Select a random song
}
\examples{
random_song()
random_song("Beyoncé")

}
